refactor: reorganize http endpoints (#602)
* fix!: fix EmbedImageStruct serialization

* fix!: Button emoji serialization

* fix!: message serialization in context

* refactor: refactor return typehints and payload naming to be consistent

* refactor: change naming of arguments

* refactor: http to folder

* Delete http.py

* http still deleted :P

* refactor: move HTTP into folder

* refactor: correct namings and imports

* refactor: add __init__, fix imports in other files

* refactor: move Guild endpoint into file; add slots

* refactor: move Member endpoint into file

* refactor: move Channel endpoint into file

* refactor: move Channel endpoint into file

* refactor: move Thread endpoint into file

* refactor: move Reaction endpoint into file

* refactor: move Sticker endpoint into file

* refactor: move interaction endpoint into file

* refactor: move webhook endpoint into file

* refactor: move emoji endpoint into file

* refactor: move emoji endpoint into file

* refactor: requested changes

* refactor: reflect breaking changes to all files

* refactor: add new HTTPClient

* docs:

* docs: fix

* refactor: rename file

* refactor: rename file

* refactor: subclassing

* refactor: remove leftover slots

* Update docs/api.http.rst


* Update api.http.rst

* Rename http.pyi to client.pyi

* Rename http.py to client.py

* Update gateway.py

* Update gateway.pyi

* Update channel.pyi

* Update guild.pyi

* Update member.pyi

* Update message.pyi

* Update role.pyi

* Update client.py

* Update client.pyi

* Update context.pyi

* Update __init__.py

* ci: correct from checks.
