+++ docs/api.http.rst
@@ -1,8 +1,68 @@
 .. currentmodule:: interactions
 
-HTTP Client
-===========
+HTTP
+====
 
-.. automodule:: interactions.api.http
+.. autoclass:: interactions.api.http.route.Route
+    :members:
+    :noindex:
+
+.. autoclass:: interactions.api.http.limiter.Limiter
+    :members:
+    :noindex:
+
+.. autoclass:: interactions.api.http.request._Request
+    :members:
+    :noindex:
+
+.. autoclass:: interactions.api.http.client.HTTPClient
+    :members:
+    :noindex:
+
+.. autoclass:: interactions.api.http.channel._ChannelRequest
+    :members:
+    :noindex:
+
+.. autoclass:: interactions.api.http.emoji._EmojiRequest
+    :members:
+    :noindex:
+
+.. autoclass:: interactions.api.http.guild._GuildRequest
+    :members:
+    :noindex:
+
+.. autoclass:: interactions.api.http.interaction._InteractionRequest
+    :members:
+    :noindex:
+
+.. autoclass:: interactions.api.http.member._MemberRequest
+    :members:
+    :noindex:
+
+.. autoclass:: interactions.api.http.message._MessageRequest
+    :members:
+    :noindex:
+
+.. autoclass:: interactions.api.http.reaction._ReactionRequest
+    :members:
+    :noindex:
+
+.. autoclass:: interactions.api.http.scheduledEvent._ScheduledEventRequest
+    :members:
+    :noindex:
+
+.. autoclass:: interactions.api.http.sticker._StickerRequest
+    :members:
+    :noindex:
+
+.. autoclass:: interactions.api.http.thread._ThreadRequest
+    :members:
+    :noindex:
+
+.. autoclass:: interactions.api.http.user._UserRequest
+    :members:
+    :noindex:
+
+.. autoclass:: interactions.api.http.webhook._WebhookRequest
     :members:
     :noindex:
