+++ interactions/api/http/__init__.py
@@ -1,0 +1,22 @@
+"""
+interactions.api.http
+
+Handles all HTTP-Requests of the library.
+"""
+
+from .channel import *  # noqa: F401 F403
+from .client import *  # noqa: F401 F403
+from .emoji import *  # noqa: F401 F403
+from .guild import *  # noqa: F401 F403
+from .interaction import *  # noqa: F401 F403
+from .limiter import *  # noqa: F401 F403
+from .member import *  # noqa: F401 F403
+from .message import *  # noqa: F401 F403
+from .reaction import *  # noqa: F401 F403
+from .request import *  # noqa: F401 F403
+from .route import *  # noqa: F401 F403
+from .scheduledEvent import *  # noqa: F401 F403
+from .sticker import *  # noqa: F401 F403
+from .thread import *  # noqa: F401 F403
+from .user import *  # noqa: F401 F403
+from .webhook import *  # noqa: F401 F403
