+++ interactions/api/http/client.pyi
@@ -1,0 +1,23 @@
+from typing import Optional, Tuple
+
+from .request import _Request
+from ...api.cache import Cache
+
+
+class HTTPClient:
+    token: str
+    _req: _Request
+    cache: Cache
+
+
+
+    def __init__(self, token: str): ...
+
+    async def get_gateway(self) -> str: ...
+    async def get_bot_gateway(self) -> Tuple[int, str]: ...
+    async def login(self) -> Optional[dict]: ...
+    async def logout(self) -> None: ...
+    @property
+    def req(self) -> _Request: ...
+    async def get_current_bot_information(self) -> dict: ...
+    async def get_current_authorisation_information(self) -> dict: ...
