+++ interactions/api/models/guild.pyi
@@ -9,7 +9,7 @@ from .misc import DictSerializerMixin, MISSING, Snowflake, Overwrite
 from .presence import PresenceActivity
 from .role import Role
 from .user import User
-from ..http import HTTPClient
+from ..http.client import HTTPClient
 
 class VerificationLevel(IntEnum):
     NONE: int
