+++ interactions/api/models/user.py
@@ -43,6 +43,9 @@ class User(DictSerializerMixin):
         "flags",
         "premium_type",
         "public_flags",
+        # TODO: document
+        "bio",
+        "banner_color",
     )
 
     def __init__(self, **kwargs):
