+++ interactions/client.pyi
@@ -4,7 +4,7 @@ from typing import Any, Callable, Coroutine, Dict, List, NoReturn, Optional, Tup
 
 from .api.cache import Cache
 from .api.gateway import WebSocketClient
-from .api.http import HTTPClient
+from .api.http.client import HTTPClient
 from .api.models.flags import Intents
 from .api.models.guild import Guild
 from .api.models.misc import MISSING, Snowflake
