+++ interactions/context.pyi
@@ -1,6 +1,6 @@
-from typing import Any, List, Optional, Union
+from typing import List, Optional, Union
 
-from .api import HTTPClient
+from .api.http.client import HTTPClient
 from .api.models.channel import Channel as Channel
 from .api.models.guild import Guild as Guild
 from .api.models.member import Member as Member
